apps/energy-engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
cvxpy>=1.4.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
